Fix a lot of GCC-specific errors and warnings.

+++ Sources/Ecc/Main.cpp
@@ -176,7 +176,7 @@ void ReplaceFileRL(const char *strOld, const char *strNew)
     }
 
     // process each charachter
-    for(int ich=0;ich<ctch;ich++)
+    for (size_t ich = 0; ich < ctch; ich++)
     {
       char *pchOld = &strOldBuff[iOldch];
       char *pchNew = &strNewBuff[ich];
